create database user

+++ backend/Models/User.cs
@@ -4,9 +4,9 @@ namespace backend.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int CPF { get; set; }
+        public long CPF { get; set; }
         public string Adress { get; set; }
         public string Email { get; set; }
-        public int Phone { get; set; }
+        public long Phone { get; set; }
     }
 }
